# openapi-v2.yaml
openapi: 3.1.0
info:
  title: eLearning Template Engine API
  description: |
    Backend API for the composable component-based eLearning authoring platform.
    Supports 89+ component types, per-interaction audio, page completion tracking,
    weighted scoring, full theming (layout + color system), and SCORM 1.2/2004 export.
  version: 2.0.0
  contact:
    name: eLearning Platform Team
servers:
  - url: /api/v1
    description: API v1 namespace
...
paths:
...
  /courses/{courseId}/discussions:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [Social]
      operationId: listDiscussions
      summary: List discussion threads
      responses:
        "200":
          description: List of discussion threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscussionThread"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Social]
      operationId: createDiscussion
      summary: Create a discussion thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussionThreadCreate"
      responses:
        "201":
          description: Thread created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscussionThread"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /courses/{courseId}/discussions/{threadId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - name: threadId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Social]
      operationId: getDiscussion
      summary: Get a discussion thread with replies
      responses:
        "200":
          description: Thread with replies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscussionThreadWithReplies"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/discussions/{threadId}/replies:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - name: threadId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Social]
      operationId: addDiscussionReply
      summary: Add a reply to a discussion thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussionReplyCreate"
      responses:
        "201":
          description: Reply created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscussionReply"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /courses/{courseId}/peer-reviews:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [Social]
      operationId: listPeerReviewSubmissions
      summary: List peer review submissions
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeerReviewSubmission"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Social]
      operationId: createPeerReviewSubmission
      summary: Create a peer review submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeerReviewSubmissionCreate"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeerReviewSubmission"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /courses/{courseId}/peer-reviews/{submissionId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - name: submissionId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Social]
      operationId: getPeerReviewSubmission
      summary: Get a submission with reviews
      responses:
        "200":
          description: Submission with reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeerReviewSubmissionWithReviews"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/peer-reviews/{submissionId}/reviews:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - name: submissionId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Social]
      operationId: addPeerReview
      summary: Add a peer review to a submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeerReviewCreate"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeerReview"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /courses/{courseId}/polls:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [Social]
      operationId: listPolls
      summary: List polls
      responses:
        "200":
          description: List of polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poll"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Social]
      operationId: createPoll
      summary: Create a poll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollCreate"
      responses:
        "201":
          description: Poll created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /courses/{courseId}/polls/{pollId}/votes:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - name: pollId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Social]
      operationId: submitPollVote
      summary: Submit a poll vote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollVoteCreate"
      responses:
        "201":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollVote"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /courses/{courseId}/polls/{pollId}/results:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - name: pollId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Social]
      operationId: getPollResults
      summary: Get poll results (tallied)
      responses:
        "200":
          description: Poll results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollResults"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/teams:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [Social]
      operationId: listTeams
      summary: List teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Social]
      operationId: createTeam
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamCreate"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /courses/{courseId}/teams/{teamId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - name: teamId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Social]
      operationId: getTeam
      summary: Get team details
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Social]
      operationId: updateTeam
      summary: Update team details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Team updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Social]
      operationId: deleteTeam
      summary: Delete a team
      responses:
        "204":
          description: Team deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Analytics & Reporting ────────────────
...
components:
...
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_-]+$"
      description: Course identifier
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict (duplicate ID)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ── Schemas ─────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ValidationErrorResponse:
      type: object
      required: [detail, errors]
      properties:
        detail:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    DiscussionThread:
      type: object
      properties:
        threadId:
          type: string
        courseId:
          type: string
        pageId:
          type: string
          nullable: true
        authorId:
          type: string
        title:
          type: string
        body:
          type: string
        isPinned:
          type: boolean
        isClosed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DiscussionThreadCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        body:
          type: string
        authorId:
          type: string
        pageId:
          type: string
          nullable: true
    DiscussionReply:
      type: object
      properties:
        replyId:
          type: string
        threadId:
          type: string
        authorId:
          type: string
        body:
          type: string
        parentReplyId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DiscussionReplyCreate:
      type: object
      required: [body]
      properties:
        body:
          type: string
        authorId:
          type: string
        parentReplyId:
          type: string
          nullable: true
    DiscussionThreadWithReplies:
      allOf:
        - $ref: "#/components/schemas/DiscussionThread"
        - type: object
          properties:
            replies:
              type: array
              items:
                $ref: "#/components/schemas/DiscussionReply"
    PeerReviewSubmission:
      type: object
      properties:
        submissionId:
          type: string
        courseId:
          type: string
        pageId:
          type: string
          nullable: true
        authorId:
          type: string
        content:
          type: object
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PeerReviewSubmissionCreate:
      type: object
      properties:
        content:
          type: object
        authorId:
          type: string
        pageId:
          type: string
          nullable: true
    PeerReview:
      type: object
      properties:
        reviewId:
          type: string
        submissionId:
          type: string
        reviewerId:
          type: string
        feedback:
          type: string
        rubricScores:
          type: object
          nullable: true
        overallScore:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PeerReviewCreate:
      type: object
      properties:
        reviewerId:
          type: string
        feedback:
          type: string
        rubricScores:
          type: object
          nullable: true
        overallScore:
          type: number
          nullable: true
    PeerReviewSubmissionWithReviews:
      allOf:
        - $ref: "#/components/schemas/PeerReviewSubmission"
        - type: object
          properties:
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/PeerReview"
    PollOption:
      type: object
      properties:
        index:
          type: integer
        text:
          type: string
    Poll:
      type: object
      properties:
        pollId:
          type: string
        courseId:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"
        allowMultiple:
          type: boolean
        isAnonymous:
          type: boolean
        isClosed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PollCreate:
      type: object
      required: [question, options]
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        allowMultiple:
          type: boolean
          default: false
        isAnonymous:
          type: boolean
          default: true
    PollVote:
      type: object
      properties:
        voteId:
          type: string
        pollId:
          type: string
        voterId:
          type: string
        selectedOptions:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
    PollVoteCreate:
      type: object
      required: [selectedOptions]
      properties:
        voterId:
          type: string
        selectedOptions:
          type: array
          items:
            type: integer
    PollResultOption:
      type: object
      properties:
        index:
          type: integer
        text:
          type: string
        votes:
          type: integer
        percentage:
          type: number
    PollResults:
      type: object
      properties:
        pollId:
          type: string
        question:
          type: string
        totalVotes:
          type: integer
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollResultOption"
        isClosed:
          type: boolean
    Team:
      type: object
      properties:
        teamId:
          type: string
        courseId:
          type: string
        name:
          type: string
        members:
          type: array
          items:
            type: string
        score:
          type: number
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TeamCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        members:
          type: array
          items:
            type: string
        metadata:
          type: object
          nullable: true
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        members:
          type: array
          items:
            type: string
        score:
          type: number
        metadata:
          type: object
          nullable: true

    # ── Analytics & Reporting ───────────────
...